rgw: unlink multipart upload parts when completing upload
Fixes: #4011
When completing the multipart upload, we also need to unlink the
parts from the bucket index. Originally we used to remove the parts
however, nowadays the parts live on as we just point the object
manifest at them. So we don't remove the objects, however, we need
to remove them from the bucket index.

Signed-off-by: Yehuda Sadeh <yehuda@inktank.com>